(BSR)[API] chore: move expired bookings functions out of scripts
